Replace asciidoc build environment with Sphinx
Prior to this commit the documentation was built with asciidoc. This
commit replaces that asciidoc build environment with Sphinx.

Note: The asciidoc documentation file `doc.asciidoc` (renamed
`index.rst` in this commit) is still asciidoc markdown. Though it
builds with Sphinx the output mostly broken. A series of commits will
follow this commit that will fix the formatting issues.

Further reading:

* http://www.sphinx-doc.org/en/1.4.8/
* http://www.sphinx-doc.org/en/1.4.8/tutorial.html